+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- Added `Quantifier` methods `#greedy?`, `#possessive?`, `#reluctant?`
 - Added `Group::Options#option_modifications`
   - shows the options enabled or disabled by the given options group
   - as with all other expressions, `#options` shows the overall active options